Add files via upload
In this post I will write my hands-on experiences about how to use an Android phone to control a robot car, which is based on an Arduino Uno board. After a couple of weeks I assembled a robot car and programmed a very simple App that can send commands to the Arduino with the help of Bluetooth. Let’s dive in!